Gitlab CI test aws access
