+++ tests/data/file_ingestion/to_ingest_1.txt
@@ -1,0 +1,3 @@
+Test 1
+Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
