+++ tests/data/file_ingestion/to_ingest_4.txt
@@ -1,0 +1,3 @@
+Test 4
+Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris
+nisi ut aliquip ex ea commodo consequat.
